Restructure file for later UI
