+++ src/Main.java
@@ -1,5 +1,7 @@
 import javafx.application.Application;
 import javafx.collections.ObservableList;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
@@ -18,6 +20,9 @@ public class Main extends Application {
         Button threeCard = new Button("Three Cards Game");
 
         HBox hbox = new HBox(5, poker, threeCard);
+        hbox.setAlignment(Pos.CENTER);
+        hbox.setPadding(new Insets(10));
+
 
         root.setCenter(hbox);
 
@@ -31,4 +36,6 @@ public class Main extends Application {
     public static void main(String[] args) {
         launch();
     }
+
+
 }
