+++ src/model/gameenum/Status.java
@@ -1,4 +1,4 @@
-package gameenum;
+package model.gameenum;
 
 public enum Status {
     READY, WAIT
