+++ src/model/gameenum/Suit.java
@@ -1,4 +1,4 @@
-package gameenum;
+package model.gameenum;
 
 public enum Suit {
     HEART(4), SPADE(1), CLUB(2), DIAMOND(3);
