+++ src/model/gamerule/Northern Poker/Rule - Vietnamese.txt
